+++ core/pom.xml
@@ -3,10 +3,10 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>com.github.andreaTP.opa.chicory</groupId>
-    <artifactId>opa-chicory-parent</artifactId>
+    <artifactId>opa-java-wasm-parent</artifactId>
     <version>999-SNAPSHOT</version>
   </parent>
-  <artifactId>opa-chicory-core</artifactId>
+  <artifactId>opa-java-wasm-core</artifactId>
   <name>Opa Wasm Chicory</name>
   <dependencies>
     <dependency>
