+++ core/src/main/java/com/styra/opa/wasm/Opa.java
@@ -1,4 +1,4 @@
-package com.github.andreaTP.opa.chicory;
+package com.styra.opa.wasm;
 
 import com.dylibso.chicory.runtime.Memory;
 import com.fasterxml.jackson.core.JsonProcessingException;
