+++ core/src/main/java/com/styra/opa/wasm/OpaAbortException.java
@@ -1,4 +1,4 @@
-package com.github.andreaTP.opa.chicory;
+package com.styra.opa.wasm;
 
 public class OpaAbortException extends RuntimeException {
