+++ core/src/main/java/com/styra/opa/wasm/OpaErrorCode.java
@@ -1,4 +1,4 @@
-package com.github.andreaTP.opa.chicory;
+package com.styra.opa.wasm;
 
 public enum OpaErrorCode {
     OPA_ERR_OK, // OPA_ERR_OK	No error.
